requirements.txt
torch~=2.0.1
uvicorn~=0.23.2
Pillow~=10.0.0
fastapi~=0.103.1
pydantic~=2.3.0
starlette~=0.27.0
torchvision~=0.15.2